# ROX-17571: Improve namespace evaluator query