--- a/central/rbac/utils/common.go
+++ b/central/rbac/utils/common.go
@@ -2,11 +2,13 @@ package utils
 
 import (
 	"context"
+	"strconv"
 
 	"github.com/stackrox/rox/central/rbac/k8srole/datastore"
 	"github.com/stackrox/rox/central/role/resources"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/sac"
+	"github.com/stackrox/rox/pkg/search"
 	"github.com/stackrox/rox/pkg/set"
 )
 
@@ -15,8 +17,8 @@ import (
 // that the role bindings were scoped to a specific namespace and referencing a cluster role.
 // If the namespace is set to empty, the given context will be used to retrieve the cluster roles.
 func getRolesForRoleBindings(ctx context.Context, roleStore datastore.DataStore,
-	bindings []*storage.K8SRoleBinding, clusterID string, namespace string) []*storage.K8SRole {
-	roleIDs, clusterRoleIDs := getRoleIDsFromBindings(bindings, namespace)
+	bindings []search.Result, clusterID string, namespace string) []*storage.K8SRole {
+	roleIDs, clusterRoleIDs := getRoleIDsFromBindings(bindings)
 	roles := make([]*storage.K8SRole, 0, roleIDs.Cardinality()+clusterRoleIDs.Cardinality())
 
 	// Fetch the roles without elevating the context.
@@ -51,17 +53,17 @@ func getRolesForRoleBindings(ctx context.Context, roleStore datastore.DataStore,
 	return roles
 }
 
-func getRoleIDsFromBindings(bindings []*storage.K8SRoleBinding, namespace string) (set.StringSet, set.StringSet) {
+func getRoleIDsFromBindings(bindingResults []search.Result) (set.StringSet, set.StringSet) {
 	roleIDs := set.NewStringSet()
 	clusterRoleIDs := set.NewStringSet()
-	for _, binding := range bindings {
-		roleID := binding.GetRoleId()
+	for _, bindingResult := range bindingResults {
+		roleID, clusterRole := convertRoleBindingSearchResult(bindingResult)
 		if roleID != "" {
 			// In cae of evaluating namespace permission where a specific namespace will be set, we will filter relevant
 			// bindings to cluster roles to the specific namespace.
 			// In case of evaluating cluster permission where no namespace will be set, we will only look at bindings
 			// without a namespace being set, since bindings bound to a specific namespace are not relevant.
-			if binding.GetClusterRole() && binding.GetNamespace() == namespace {
+			if clusterRole {
 				clusterRoleIDs.Add(roleID)
 			} else {
 				roleIDs.Add(roleID)
@@ -80,4 +82,20 @@ func fetchRoles(ctx context.Context, roleStore datastore.DataStore, roleIDs set.
 		}
 	}
 	return roles
+}
+
+func convertRoleBindingSearchResult(r search.Result) (string, bool) {
+	var (
+		roleID      string
+		clusterRole bool
+	)
+	roleIDMatches := r.Matches["k8srolebinding.role_id"]
+	if len(roleIDMatches) == 1 {
+		roleID = roleIDMatches[0]
+	}
+	clusterRoleMatches := r.Matches["k8srolebinding.cluster_role"]
+	if len(clusterRoleMatches) == 1 {
+		clusterRole, _ = strconv.ParseBool(clusterRoleMatches[0])
+	}
+	return roleID, clusterRole
 }