--- a/central/rbac/utils/namespace_permission_evaluator.go
+++ b/central/rbac/utils/namespace_permission_evaluator.go
@@ -47,18 +47,26 @@ func (c *namespacePermissionEvaluator) RolesForSubject(ctx context.Context, subj
 
 func (c *namespacePermissionEvaluator) getBindingsAndRoles(ctx context.Context, subject *storage.Subject) ([]*storage.K8SRoleBinding, []*storage.K8SRole) {
 	q := search.NewQueryBuilder().
+		AddStringsHighlighted(search.RoleID, search.WildcardString).
+		AddBoolsHighlighted(search.ClusterRole, true).
+		AddBoolsHighlighted(search.ClusterRole, false).
 		AddExactMatches(search.ClusterID, c.clusterID).
 		AddExactMatches(search.Namespace, c.namespace).
-		AddLinkedFields(
-			[]search.FieldLabel{search.SubjectName, search.SubjectKind},
-			[]string{search.ExactMatchString(subject.GetName()), search.ExactMatchString(subject.GetKind().String())}).
+		AddExactMatches(search.SubjectName, subject.GetName()).
+		AddExactMatches(search.SubjectKind, subject.GetKind().String()).
 		ProtoQuery()
-	roleBindings, err := c.bindingsStore.SearchRawRoleBindings(ctx, q)
 
+	roleBindingsSearchResult, err := c.bindingsStore.Search(ctx, q)
 	if err != nil {
-		log.Errorf("error searching for roleBindings: %v", err)
+		log.Errorf("Error searching for roleBindings: %v", err)
 		return nil, nil
 	}
-	roles := getRolesForRoleBindings(ctx, c.roleStore, roleBindings, c.clusterID, c.namespace)
-	return roleBindings, roles
+
+	roles := getRolesForRoleBindings(ctx, c.roleStore, roleBindingsSearchResult, c.clusterID, c.namespace)
+	bindings, _, err := c.bindingsStore.GetManyRoleBindings(ctx, search.ResultsToIDs(roleBindingsSearchResult))
+	if err != nil {
+		log.Errorf("Error retrieving role bindings: %v", err)
+		return nil, nil
+	}
+	return bindings, roles
 }