## Description

There has been a notion that sometimes the queries for the namespace evaluator time out within CI.

The main issue is that we previously when searching the K8S role bindings were taking the full object into account, whilst for retrieving the correct set of roles only the `cluster_role` and `role_id` search fields were required.

The query has been adjusted and instead of using `SearchRawRoleBindings` we are now using `Search`.

However, at the end, we will still be required to pull the full object for evaluating the RBAC for specific subjects due to needing the full `Subject` field from the role binding. While this adds a little bit of overhead, we do this now separately instead of within the search.

While working on the benchmark tests, took liberty to rewrite the existing tests from using mocks to SQL integration tests.

## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps
- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))

If any of these don't apply, please comment below.

## Testing Performed

- see CI.

- Benchmark comparison with `benchstat` from master and after the changes to the queries:
```
benchstat old.txt new.txt                                                                                                                                                                                                                                                                                             
pkg: github.com/stackrox/rox/central/rbac/utils
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                                  │    old.txt    │               new.txt               │
                                                  │    sec/op     │   sec/op     vs base                │
GetBindingsAndRoles/run_get_bindings_and_roles-12   3182.1µ ± 11%   907.6µ ± 3%  -71.48% (p=0.000 n=10)

```